fix(vmm): Add vendor ID normalization
The custom CPU template feature allows users to modify vendor ID. This
results in a situation where the vendor ID mismatch the brand string,
because the normalization is enforced on the brand string based on the
host value.

Adds additional normalization on the vendor ID to prevent the above
situation.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>