test: do not load baselines into global constant
Loading baselines into a global constant causes pytest test collection
to fail if run on a platform for which no baselines exist. This is the
case even if the '-m nonci' flag (which enables the tests in question)
is NOT passed, making running tests on these platforms quite difficult.

Do not do this for the block performance test yet, as it still depends
on the baseline file for its parameterization, and changing this
requires baseline regathering.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>