ci: publish individual datapoints in metrics
Currently performance tests publish an aggregate of metrics over their
whole duration, like an average or a percentile.

This change: 1) allow publishing of "raw" datapoints, so we can do
different aggregations and better understand the data, and 2) rework the
test_network_latency test to output them, since it is the simplest one.

Signed-off-by: Pablo Barbáchano <pablob@amazon.com>