bench: remove snapshot benchmark from vmm crate
The benchmark in question was booting up microvms and interacting with
the file system. These operations go fundamentally against the spirit of
a microbenchmark, as they add significant variability (which was
observable in the new criterion-based benchmark test, which reported
that running the benchmark twice in succession with no code changes
yielded statistically significant different measurements). Additionally,
the benchmark in question is not part of our current testing suite, and
the important part of it is covered by the benchmark in the snapshot
crate).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>