test: change the echo server in test_5_snapshots
With AL2023 and with a 4.14 kernel and some machine configurations, this
test can fail due to the vsock clients timing out during connect.

The problem disappears if we run socat instead of the internal echo
server that we run in a thread with Python. It may be the echo server is
buggy or the Python GIL is acting up.

So, replace the single-threaded echo server with socat, which can fork
per client and is an external process.

Signed-off-by: Pablo Barbáchano <pablob@amazon.com>