test: handle perf retries at test level instead of buildkite
Doing the retries at the buildkite level caused a single individual test
instance failure to retry ALL tests in the same file, including those
that were passing. This is a waste of time and compute resources, and
additionally increases the probability of spurious failures, as two
different tests could spuriously fail in subsequent retries.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>