Merge vmm build script into firecracker build script
The build script invokes cargo, which breaks tooling such as miri when
attempting to run it on the vmm crate. With us currently moving all
other crates into a single library crate, this is starting to become
annoying.

This also moves the code that reads the compiled seccomp config and
includes it into the firecracker binary. Outside the firecracker binary
crate, the code in question was only used in unittests. Unittests that
used the production filters now use empty filters. This is okay because
we run our integration test suite with the production filters, and the
filters are out of scope for unittests anyway.

With moving the code to a binary crate, rustc's dead code analysis
kicked in, so some error variants are removed.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>