aarch64: enable PAC support if present
ARMv8.3 introduces support for Pointer Authentication Code (PAC). PAC is
a defensive mechanism against attacks like buffer overflows. This commit
enables Firecracker to check the host for PAC support and if it is
supported expose the capability in the guest VM.

This breaks backwards compatibility of snapshots. We cannot build
snapshots targeting versions < 1.4.0.

Signed-off-by: Babis Chalios <bchalios@amazon.es>