test: change way api_server timings are determined
Previously, when checking that api server requests complete in under
700ms, we measured this client-side, which included whatever latency was
added by python/the network stack.

Now we instead directly utilize the timing information firecracker
provides in its logs in our integration tests (a log message for this
was previously present, and has been upgraded to info level so that we
can access it in the integration tests).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>