fix: panic when calling /cpu-config with log level set to debug
Due to use of the log_enabled! macro to determine the verbosity of a log
message, a call to Logger::enabled(...) was made by the logging
framework, if log verbosity was at least Debug (the macro first checks
if the global log level allows Debug, and if yes, calls to Log::enabled
to see if there are any custom constraints by the `Log` instance). As
our implementation of `Log::enabled` was simply `unimplemented!()`, we
got a panic.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>