balloon: don't map over existing regions
Firecracker used to map over the mmap when it was a restore. For
CodeSandbox this is not needed, as we use UFFD and we don't have the
limitation that the map was used for.

In fact, this is problematic for the UFFD handler, because it implicitly
unregisters the UFFD handler from the reclaimed range.