virtio: block: fix integer overflow
When computing the end of the destination slice for a configuration
write in Block::write_config(), an addition might overflow. This will
immediately cause a panic in debug builds, and panic on slice
indexing a few lines below for release builds.

Fix this by using `usize::checked_add()` and `slice::get_mut()`.

Signed-off-by: Carlos López <carlos.lopez@suse.com>