virtio: net: fix integer overflow
Fix a potential integer overflow when computing the end of the
destination slice during a configuration write in
Net::write_config(). This bug leads to a panic when adding with
overflow or indexing the slice a few lines below, depending on the
build.

Fix this by using `usize::checked_add()` and `slice::get_mut()`.

Signed-off-by: Carlos López <carlos.lopez@suse.com>