fix(test) remove check for v0.23.0 snapshot
test_snapshot_compatibility virtio-balloon integration test was trying
to ensure we cannot create a snapshot from a microVM with a balloon
device targeting Firecracker v0.23.0. The test expects to find an API
response mentioning that the target Firecracker version does not support
balloon. Instead, at times it receives an error due to the fact that the
"target Firecracker version does not support notification suppression".

The latter error message is emitted when we try to create a snapshot
targeting Firecracker version earlier 1.0.0, since notification
suppression support was introduced with 1.0.0.

In fact, the virtio-balloon specific check should never be triggered,
only the notification suppression one. The test was working was by
chance. The notification suppression condition was not triggered because
we were trying to take a snapshot immediately after booting the VM,
before the block device was initialized.

This commit removes the check all together, since v0.23.0 is not
supported any more.

Signed-off-by: Babis Chalios <bchalios@amazon.es>