test(perf): Do no parameterize restore tests by snapshot_type
Testing DIFF snapshots for the restore test is not an interesting test
permutation, as creating a DIFF vs FULL snapshot only has an impact on
the creation. Before a diff snapshot can be restored, it has to be
rebased, at which point the result is a full snapshot. So the CI tests
related to testing latency of restore with DIFF snapshots do not make
much sense (admittedly, there is one extra valloc call in the KVM
source [1], as dirty-page tracking will be enabled on the load path, but
this hardly warrants an entire new test permutation).

[1]: https://github.com/torvalds/linux/blob/a901a3568fd26ca9c4a82d8bc5ed5b3ed844d451/virt/kvm/kvm_main.c#L1626-L1638

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>