test: Drops unimplemented MSRs
msr_reader.sh queries some ranges of MSR indices regardless of whether
they are implemented or not, and it has saved them with the STATUS field
("implemented" or "unimplemented") so far. However, a large portion of
them are unimplemented, resulting in large numbers of lines of baseline
files.

This commit drops unimplemented MSRs from msr_reader.sh. This should be
fine since not having some MSR indices implies that they are
unimplemented.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>