test: Name performance test data consistently
Changes the dimensions of the raw metrics output by the performance
tests to match their aggregates. Prior to this change, all
parameterizations of a single tests were only tagged with the test
function name. This change instead includes the "tag" as well as full
parameterized pytest test name, just like the aggregates.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>