test(perf): Aggregate snapshot latencies to average
The statistical test used by firecracker performance tests is a
simplified t-test. This is based on the central limit theorem, and is
thus only valid for testing _averages_. Nevertheless, the snapshot tests
tried to use it for detecting deviations of P90/P50 values. This is
unsound, and the results are most definitely not what was intended. This
commit changes the aggregate collected from the snapshot tests to be an
average instead, to make the test sound.

Note that since we also emit the individual data points, we can always
reconstruct percentiles.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>