Use 0x40000010 CPUID leaf to announce frequencies
The 0x40000010 CPUID leaf is used to announce TSC and local APIC
frequencies to VM guests; this allows them to boot faster by bypassing
clock calibration loops.

When resuming from a snapshot the TSC/APIC leaf will exist or not
depending on whether it was created in the original VM, since we
store CPUID state as part of the snapshot.

Signed-off-by: Colin Percival <cperciva@freebsd.org>