test: keep a copy of test run
This change makes it so we always generate a test report, not only
while performing a release.

This is useful for example to analyze the behavior of a test
historically. We already have metrics with some metadata that include a
lot of the information here, but these reports also include
stdout/stderr and log messages, that could be useful to see in
retrospect.

More info at https://pypi.org/project/pytest-json-report/

Signed-off-by: Pablo Barbáchano <pablob@amazon.com>