test: add a test to measure Firecracker memory overhead
Our current binary size test measures space in disk, as a proxy for used
memory. This is not accurate, as parts of the binary are not loaded into
memory.

This commits adds a new test that measures memory overhead directly and
published metrics for different guest configurations.

Also remove thresholds from test_binary_size.py test, as we measure
memory overhead directly now.

Signed-off-by: Pablo Barbáchano <pablob@amazon.com>