fix(vmm): Change T2A to set EferLmsleUnsupported to 1
CPUID.0x80000008:EBX[20] (EferLmsleUnsupported) indicates that
EFER[LMSLE] is not supported. As described in AMD64 APM vol.2,
EFER[LMSLE] has already been deprecated and is not suported by all
processor implementations (including m6a.metal).

KVM allows nested virtualization guests to write EFER[LMSLE] only for
very specific software (SLES11 version of Xen 4.0 to boot nested SVM).
On non-nested virtualized guests, writing to EFER[LMSLE] generates #GP
which is identical to behavior on CPUs where the EferLmsleUnsupported
bit is set to 1. As T2A template disables SVM, it is safer and better to
let guests know that writing to EFER[LMSLE] is not supported.

As a side note, to revert the upstream commit allowing it on nested
virtualized guests, some discussions were made but was cut off in the
middle.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>