fix: specify resolver in Cargo.toml
With one of the recent rust toolchain upgrades, we started getting a
warning about different defaults for the results (workspaces default to
resolver = "1", yet individual crates default to resolver = "2" since
edition 2021). To fix this warning, explicitly set the resolver to
version 2 in our top level Cargo.toml

The change in resolvers is related to how features are merged. Instead
of unifying them (meaning that features enabled in dependencies "leak"
into other crates), they are now kept disjoint. This is why the regex
dev-dependency now needs the `unicode-perl` feature explicitly enabled.

[1]: https://doc.rust-lang.org/edition-guide/rust-2021/default-cargo-resolver.html

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>