test: Adjust tests to uphold debug_assert!s
Functions in queue.rs now have debug_assert!s that validate they are
only executed if the queue's in-memory layout is valid (to match
assumptions made in our kani proofs). Two unittests were not respecting
this requirement. This commit updates one of them to correctly
initializes queues, and removes the other (as it was explicitly testing
an operation on a mis-aligned queue).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>