Fix: During normalize, ignore absence of frequency in host brand string
As reported in issue #3995, some instances might not have
frequency mentioned in the host model name.
Though it is not essential for us to include host frequency in
the normalized brandstring we expected it to be present since
our internal instances had frequency present in the model name.
Given that checking frequency or other checks in host brand string
are not always valid for external instances, use a default string if
any check related to host brand string fail.

Signed-off-by: Sudan Landge <sudanl@amazon.com>