kani: proof-harnesses for IoVecBuffer(Mut)
IoVecBuffer(Mut) are wrappers on top of guest buffers that not
necessarily contiguous in memory. The types provide methods for
reading/writing from/to these memory buffers.

Memory accesses to buffers provided by the guest is not safe (in terms
of Rust safety). For example, we cannot create slices on top of the
underlying addresses because we cannot guarantee that the guest will
behave in the way Rust expects it to do. As a result, IoVecBuffer(Mut)
describe the underlying memory using raw pointers and implement
read/write methods on them using unsafe primitives.

This commit adds proof-harnesses for these read/write methods.

Signed-off-by: Babis Chalios <bchalios@amazon.es>