Fix: Avoid microvm build if Api server thread fails
PrebootApiController::build_microvm_from_requests() should ideally
start after Api server thread was able bind to a socket.
Not doing so results in intermittent failures with below message:
```
The channel's sending half was disconnected. \
Cannot receive data.: RecvError
```
This message makes more sense if API server thread was actually able
to bind to a socket and then disconnected.
So, change the code to make sure build_microvm_from_requests is
called only if API server thread is able to bind to a socket.

Signed-off-by: Sudan Landge <sudanl@amazon.com>