feat(vmm): add block device subtypes list in device state
In order to be able to restore different block device subtypes
from a snapshot, this introduces an additional list
block_device_subtypes in the DeviceStates struct.

This approach was chosen to minimise disturbance
of the versionized DeviceStates struct, compared to
an alternative solution of converting ConnectedBlockState
into an enum.

On save, if the snapshot version does not support device subtypes,
we only proceed with taking the snapshot if only SUBTYPE_BLOCK_FILE
devices are attached to the microVM.

On restore, if the snapshot version does not support device subtypes,
we populate the list with SUBTYPE_BLOCK_FILE entries on restore.

Signed-off-by: Nikita Kalyazin <kalyazin@amazon.com>