test: Make test_guest_cpu_config_change optional
[ Upstream commit 554635c553555deffb58b59d9de1858d65d2695c ]

This test compares the baseline guest_cpu_config gathered in the past
with the current one and fails potentially due to firecracker's code
changes or kernel/microcode updates. To be aware correctly that CPU
config has been changed by firecracker, the test was set as a mandatory
requirement to merge PRs.

However, since the introduction of the test (commit 8161c0bc
("test(tool): verify CPU config has not changed") on 2023 Apr 5, the
following CPU config changes have been caused by non-firecracker
changes:
- The RRSBA bit of IA32_ARCH_CAPABILITIES MSR flipped from 0 to 1 on
  Intel Cascade Lake.
- The GDS_CTRL bit of IA32_ARCH_CAPABILITIES MSR flipped from 0 to 1 on
  combinations of (Intel Skylake, Intel Cascade Lake, Intel Ice Lake) x
  (host kernel 4.14).

Although this kind of CPU config changes should be investigated and
handled appropriately, it blocks all PRs and we cannot merge any PRs
until it is fixed. So this commit makes the test optional.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>