test: Upload coverage data to codecov.io
Instead of doing a baseline-based coverage test, upload to codecov.io to
get their GitHub integration to post a detailed comment about coverage
changes, as well as to utilize their detailed coverage reports.

The token needs to be accessible inside of docker for us to be able to
upload coverage reports, so we feed it through.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>