fix: handle BUILDKITE_PULL_REQUEST correctly for non PRs
For pull requests, the variable is set to the number of the pull
request. For builds triggers not on pull requests its... the string
literal "false". That broke submitting coverage data from non-pr
branches because it tried to submit coverage data for a pull request
with number "false", instead of submitting for the branch on which the
test was run.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>