test: Prevent optional pipeline from running nonci tests
We have recently introduced tests that are marked with the combination
of nonci and optional. This is because the tests are parameterized
to use different templates on x86 and ARM, but the ARM template tests
cannot be run in our PR CI due to requiring a specially compiled host
kernel. Therefore, whatever can run in our CI should run in optional,
and ARM should only run nightly.

This commit allows for what is described above.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>