Miscellaneous fixes that don't fall anywhere else
In vmm/builder.rs, the CPU count can be fetched from the VM
configuration without having to do a type conversion.

In the serial device and the networking code, I added explicit checks to
ensure that values fit in bounds.

In the vCPU code, the existing code uses floats for tolerance values. We
can replace these all with integer multiplications and divisions.

Signed-off-by: Jonathan Browne <12983479+JBYoshi@users.noreply.github.com>