Add try_from to usages of generated constants
Some of the constants used in Firecracker are automatically generated
from C bindings. C's #define constansts do not specify types, so they
are all set to the same type, and we can't change this behavior in the
generated code. To work around it, I use try_from() for all of these
cases.

An exception is when defining enum values in terms of C constants.
Currently, using try_from() or try_into() in const values is unstable,
so I continue to use "as" for those conversions and mark it as
explicitly allowed.

Signed-off-by: Jonathan Browne <12983479+JBYoshi@users.noreply.github.com>