test: rewrite the memory monitor
The memory monitor is a subsystem that monitors Firecracker's RSS
doesn't go above 5MB. It is implemented as a thread that shells out to
pmap to get the RSS every 50ms. It is intrussive and brittle as it
sometimes break when running tests in parallel.

Rewrite the monitor to use psutil.

Signed-off-by: Pablo Barbáchano <pablob@amazon.com>