[Test code] Shrink the sizes of some constants
This commit changes several constants to use smaller types. This means
we don't have to downcast them to use them, and we can safely convert
them to larger types when needed.

One limitation with this is that Kani doesn't support casts in length
fields, so I changed those cases to use a separate constant containing
the cast.

Signed-off-by: Jonathan Browne <12983479+JBYoshi@users.noreply.github.com>