test: Do not use Intel DB inside guests
spectre-meltdown-checker.sh uses by default a builtin Intel DB that
holds information about which processors are affected by which
vulnerabilities and require microcode updates. However, since KVM and
firecracker modifies CPUID and MSRs of guests, the DB does not make
sense inside guests. This commit makes the test using the script not use
the DB and evaluate whether vulnerable or not based on CPUID and MSRs
visible inside guests. This change helps guests with GDS_NO = 1 pass the
test when the required KVM patch is applied.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>