test: Add a snapshot restore test compatible with A/B-testing
We add a version of the nightly snapshot restore test that does not
interface with the current baseline-based testing framework. Instead, it
will only write the latency samples to test_results. It respects the
--binary-dir option, meaning this test can be used to collect latency
samples for firecracker binaries compiled from old revisions.

We choose this approach of doing a "data production only" test, instead
of having the actual A/B-test be done inside of pytest for multiple
reasons:
- We cannot compile old firecracker versions from inside pytest, as this
  would require us to nest docker (or rely on the old firecracker
  revision being compilable with the current docker container).
- Doing the A/B-orchestration outside of the test means the test does
  not need to support "metrics only" and "A/B" modes (with the former
  being required for nightly data collection runs).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>