vmm: Show warning message on processors prior to Cascade Lake with C3
As mentioned in the previous commit, the kernel does not apply the
mitigation against MMIO stale data vulnerabiliies when it is running on
a processor that does not enumerate FBSDP_NO, PSDP_NO and SBDR_SSDP_NO
on IA32_ARCH_CAPABILITIES MSR. This is the case on Intel Skylake on our
test platform. To let users know it, this commit changes to show a
warning message and adds descrption into the CPU template doc.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>