test(uffd): serve a single page per fault in valid handler
With kernels >= 5.11, Linux allows a process that creates a userfaultfd
file descriptor to opt out from handling kernel triggered page faults.
`userfaultfd-rs` by defaults does this, so in 6.1 we will not be able to
handle page faults until we fix this option.

Our current valid page fault handler, will not exercise the issue
because it faults in the entire memory the first time a page fault
happens. In order to allow catching such errors, we change the page
fault handler to, instead, serve a single page at a time. This is slow,
but we are just testing here.

Signed-off-by: Babis Chalios <bchalios@amazon.es>