test(uffd): refactor tests to print UFFD handler logs
The way we were checking if a UFFD handler was alive in tests involved
us sending some bytes to it over stdin and waiting to read something
over stdout with success. To achieve that we need to "pipe" the
handler's stdin/stdout, which means that we can't any more peep in its
output.

This commit changes the mechanism to use `Popen.poll()` to ensure that
the process is still alive and keep track of its stdout and stderr in a
file.

Signed-off-by: Babis Chalios <bchalios@amazon.es>