uffd: opt-in to handle kernel triggered page faults
Since 5.11 Linux allows a process to choose if it wants to handle page
faults that were triggered in kernel space. userfaultfd-rs by default
chooses not to. Even though, this option does not affect kernel 5.10, it
means that in 6.1 we will not be able to handle such page faults.

This commit explicitly asks to handle kernel-triggered page faults.
Moreover, it updates the userfaultfd-rs version to a newer one that
uses the /dev/userfaultfd API to create file descriptors. This API does
not require a process to have `CAP_SYS_PTRACE` capability in order to
create a file descriptor capable of handling kernel-triggered page
faults.

Signed-off-by: Babis Chalios <bchalios@amazon.es>