iovec: improve way or reading/writing subranges
We have a IoVecSubregion type for describing sub-regions inside an
initial IoVecBuffer(Mut) type. This is useful type for de-duplicating
the code that finds a valid sub-region so that we can do something with
it later. However, the type was essentially duplicating (some of) the
state of the initial IoVecBuffer(Mut). The memory overhead should not be
significant, but the book-keeping overhead (creating yet another Vec)
is noticeable. Moreover, the interface was leading us to have some
nested loops. This was increasing significantly the time for kani
harnesses.

This commit drops the IoVecSubregion type and replaces the functionality
with a simple loop that performs the necessary operation on a valid
sub-range.

Signed-off-by: Babis Chalios <bchalios@amazon.es>