[Test code] Apply bit masks to allow some wrapping
In these cases, values are intentionally truncated, either to split them
into multiple values or to ensure that user-supplied data does not cause
Firecracker to panic with a try_from/unwrap failure.

Clippy will allow us to use "as" when we use bit masks to limit the size
of an explicitly-sized value (u16, u32, u64). It doesn't work on usize
values or on some more complicated expressions.

Signed-off-by: Jonathan Browne <12983479+JBYoshi@users.noreply.github.com>