Fix spelling of 'nomodule' kernel param
Changes "nomodules" in the default kernel cmdline to "nomodule",
the correct spelling. The old argument was not recognized by the
kernel and incorrectly forwarded to init binaries as argv.

Note, this is a breaking change for setups that use the default guest
kernel command line and also happen to be loading kernel modules at
runtime. This may also break init binaries that were inadvertently
relying on the incorrect "nomodules" arg to be present in argv.

Signed-off-by: Brandon Duffany <brandon@buildbuddy.io>