test: test the current binary as a release
With our current firecracker_release fixture, some tests are only run
against released firecracker versions that we keep in an S3 bucket. This
means the binary in the working copy does not execute those tests, and
we may only realize until we make a release, at which point is too late
and we have to make a patch release.

Fix it by extending the firecracker_release fixture to also test the
current binary as a regular release, so it goes through the same tests.

Signed-off-by: Pablo Barbáchano <pablob@amazon.com>