ARM64: Add bounds checks on device reads/writes
Errors here appear to just emit warnings, so I added a check here to
ensure the offset is in bounds.

I've done this in a separate commit because my dev setup doesn't use
ARM. In case something ARM-specific breaks in CI, I'd like to be able to
keep those changes in their own commits so I can debug more easily.

Signed-off-by: Jonathan Browne <12983479+JBYoshi@users.noreply.github.com>