test(jailer): remove test_negative_file_size_limit
The test sets RLIMIT_FSIZE to 2M and expects
for the jailer to kill Firecracker
when the latter tries to create a big memory snapshot file.

This test is no longer relevant, because Firecracker
now creates guest memory with memfd, which makes minimum
viable RLIMIT_FSIZE equal to the guest memory size.

Signed-off-by: Nikita Kalyazin <kalyazin@amazon.com>