fix: Avoid reallocating fd table during snapshot restore
Linux maintains a per-process file descriptor table. By default, this
table has size 64. Whenever space in this table runs out, the file
descriptor table gets reallocated, with its size increased to the next
power of two.

Firecracker creates a lot of eventfds and timerfds for its devices.
These are created in the hot path of snapshot restore. For medium to
large microVMs, firecracker uses more than 64 file descriptors, meaning
we reallocate the file descriptor table on the hot path. However, this
reallocation uses a RCU, meaning the kernel needs to hold a lock. To
acquire this lock, it needs to wait for all other accesses to the file
descriptors to cease, which introduces a severe latency to
snapshot-restore times (between 30ms and 70ms).

This commit avoids that latency by ensuring the file descriptor table is
large enough to hold the jailer-defined limit of file descriptors at
process start already. This avoids reallocating the file descriptor
table at runtime (and the memory overhead is negligable, as each entry
in the fdtable is simply a pointer).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>