cleanup: Remove now unneeded wrapper types from mptable.rs
Before e0c9ac93788d0b16ddf76a8a5ef2e2fe208297eb we needed a workaround
to be able to implement the ByteValued trait for the various mpspec
structures, as they were defined in a different crate (and that crate
contained autogenerated bindings). Thus we introduced wrapper types to
work around the orphan rules.

Now however all code lives in the vmm crate, meaning the wrapper types
are no longer neccessary.

Furthermore, we do not need to use `unsafe` in compute_checksum, as it
is only called for types that implement `ByteValued`.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>