test: Do not have git_ab_test chdir into temporary checkout
Instead pass the path to the temporary checkout to the test_runner
closure. The reason being that by automatically chdir-ing into the
temporary directory, it is difficult for the test_runner to get access
to things outside of the temp dir (which has to be explicitly passed via
closure). Without, the test runner knows both where the permanent
firecracker checkout is (via Path.cwd()), and can easily chdir to the
temporary one via the ab_test.chdir() context manager.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>