snapshot: deprecate forward compatibility
Forward compatibility refers to the ability of Firecracker to create
snapshots for older version formats. This feature is now deprecated and
it will be removed in a subsequent release. Snapshot support is in
developer preview, so dropping support for it does not require us to
increase Firecracker's major version.

In practice, deprecating this feature means that we deprecate the
`version` field of the `PUT /snapshot/create` request. This commit, adds
the "deprecated" header in the response of the endpoint if `version` is
defined and documents the deprecation in CHANGELOG and snapshot
documentation.

Signed-off-by: Babis Chalios <bchalios@amazon.es>