test(vsock): avoid running multiple guest echo servers
In negative vsock tests, a guest echo server
was started both independently and inside
check_host_connections() leading to multiple
instances running in the guest at the same time.

This commit removes starting the server inside
check_host_connections(). This makes it inconsistent
with check_guest_connections() that still starts
a host echo server internally, but removing it
is not currenlty warrranted.

Signed-off-by: Nikita Kalyazin <kalyazin@amazon.com>