test(vsock): use default blob size in epipe test
A large blob size (20M) is used in the first part of the test
to increase likelihood of triggering the epipe condition.
In the second part of the test, a call to
check_host_connections() is performed to verify that Firecracker
can still operate after receiving a sigpipe.
There is no need to use a large blob size in the second
part.

This commit gets back to the default blob size
in the second part of the test to reduce the test duration.

Signed-off-by: Nikita Kalyazin <kalyazin@amazon.com>