vmm: Deprecate static CPU templates
Deprecate the `cpu_template` field in PUT and PATCH reqeusts on the
`/machine-config` API by following the runbook for API changes.

To differentiate between explicitly specifying `None` and not specifying
anything, wrap the `cpu_template` of the `MachineConfig` with `Option`.
With this, it can notify the deprecation of only users using the field.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>