feat: add a debug log for vm build time
Add a couple of events to the debug logs. With this information we can
extract information of how long did those events take.

Added as debug messages to make it clear to customers that these are
unstable and may change format or disappear at any moment, at should not
be depended upon.

Signed-off-by: Pablo Barbáchano <pablob@amazon.com>