fix: Add test to check static linking of binary
host.get_binary() today returns path of a binary which is
compiled with the --release flag and because we use default
target as musl this also means that today we this function
returns a statically linked release profile binary.
Since we use this binary in our CI add a test to check if
the Firecracker binary is actually compiled using musl
(static link) or gnu (dynamic link).
We use `file` command because it is already available in devctr.
Note: This test doesn't guarantee that we use a statically linked
binary in our Firecracker release, the commit after this handles that.

Signed-off-by: Sudan Landge <sudanl@amazon.com>