fix:create dummy test-report for release.sh sanity test
Release process involves running the command `devtool -y make_release`
which runs a bunch of tests first and then runs the release.sh script.
The tests create a report at tests/test-report.json which is copied by
release.sh to a release folder. The release.sh does not check contents
of the report and simply copies it.
However, when release.sh is run as part of buildkite's release sanity
build no other tests are run and so test-report.json is not available
resulting in below error:
```
cp: cannot stat 'tests/test-report.json': No such file or directory
```
So, create a dummy tests/test-report that release.sh can copy and
avoid above the error.

Signed-off-by: Sudan Landge <sudanl@amazon.com>