test: Add GitHub action for triggering A/B-Tests
The Action triggers on push events to main instead of on pull request
merge events. This is because pull request events are triggered in the
context of a fork, whereas push events are triggered in the context of
the firecracker repository. Only the latter context allows access to
GitHub secrets.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>