refactor: move dumbo, mmds and logger to vmm (9/11)
Now that we have moved logger to vmm unit tests
in vmm_config related to init metrics fail because
they refer to the global METRICS which is already
initialized by the unit tests in logger::metrics.
So we fix the test by not using global METRICS in
logger::metrics.

Note: This is part of commits having the same title.

Signed-off-by: Sudan Landge <sudanl@amazon.com>