ARM64: Update types for (in)equality checks
For equalities and inequalities, we can change the types on each side to
always convert to the larger type, which ensures all conversions are
safe.

I've done this in a separate commit because my dev setup doesn't use
ARM. In case something ARM-specific breaks in CI, I'd like to be able to
keep those changes in their own commit so I can debug more easily.

Signed-off-by: Jonathan Browne <12983479+JBYoshi@users.noreply.github.com>