test: Add block performance A/B-test
This test runs a fio workload and emits read/write throughput metrics to
cloudwatch in a way that they will get picked up by `ab_test.py`.

Timeout adjusted to match runtime of a typical A/B block test.

Since we remove the `env_id` strings, we change the path to which we SCP
the fio output to simply be `chroot/fio_output`. Since every test runs
in its own microVM, there is no need for these paths to be unique
between tests.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>