fix: convert `numpy.int64` to float in `ab_test.py`
Apparently, aws_embedded_metrics understands what a numpy.float64 is,
but a numpy.int64 is the devil's work and causes a JSON serialization
exception. We get these sometimes if the mean difference is of
significant magnitude (in which case rounding happens). Thus, cast
everything to float to get aws_embedded_metrics to understand it.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>