test: Have `ab_test.py` output correct units
Previously it always interpreted the magnitude of the regression as
being milliseconds. It now correctly determines the unit from the EMF
logs and pretty-prints it (e.g. no more trying to have a guess at how
many MB are in 132432Bytes, it will reduce it to the most readable unit
automatically).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>