test: A/B-ify test_cargo_audit
The test_cargo_audit test can fail due to new advisories being published
on crates that firecracker already consumes. We already get notified of
this because we do a nightly run that alerts us whenever any advisory
for one of our dependencies exists. There is no need to do the same on
PRs, as that means all development activities are blocked until the
advisory is resolved (which might be non-trivial). Thus use A/B-testing
in CI to only fail a PR if a newly introduced dependency has a
pre-existing advisory (in which case we do not want to introduce the
dependency). This is done by computing the set of
warnings/vulnerabilities that currently affect the repository, and
ensuring that there are no additions to this set (while subtractions,
for example due to a PR removing a vulnerable dependency, are allowed).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>