test: use ip netns exec to ssh
Use `ip netns exec` when calling the ssh process. This means the pytest
thread does not have to switch network namespace and back.

We already use the same approach just below for the tap device, so now
there is a single way to enter a network namespace.

Also drop the nsenter Python package dependency, which seems
unmaintained.

Signed-off-by: Pablo Barbáchano <pablob@amazon.com>